requirements.txt
bidict==0.21.3
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
dataclasses-json==0.5.6
Flask==2.0.2
Flask-SocketIO==5.1.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
marshmallow==3.14.0
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
python-dotenv==0.19.1
python-engineio==4.2.1
python-socketio==5.4.1
requests==2.26.0
SQLAlchemy==1.4.26
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.7
Werkzeug==2.0.2
